Extracted LogicalXORFitnessService
