+++ example/logical_xor/lib/logical_xor_fitness_service.dart
@@ -1,0 +1,95 @@
+import 'dart:math';
+
+import 'package:genetically_evolving_neural_network/genetically_evolving_neural_network.dart';
+
+/// This fitness service will be used to score a logical XOR calculator. The
+/// output should only return true if a single value is 1.0 and both other
+/// values are 0.0. There should be one exclusive positive value! The more
+/// correct guesses that a NeuralNetwork makes, the higher its fitness score
+/// will be.
+class LogicalXORFitnessService extends GENNFitnessService {
+  /// The list of logical inputs for your XOR calculator. The possible options
+  /// are 0 or 1, effectively true or false.
+  List<List<double>> logicalInputsList = [
+    [0.0, 0.0, 0.0],
+    [0.0, 0.0, 1.0],
+    [0.0, 1.0, 0.0],
+    [0.0, 1.0, 1.0],
+    [1.0, 0.0, 0.0],
+    [1.0, 0.0, 1.0],
+    [1.0, 1.0, 0.0],
+    [1.0, 1.0, 1.0],
+  ];
+
+  /// The list of logical outputs for your XOR calculator. These are the
+  /// expected outputs respective to the logical inputs.
+  List<List<double>> targetOutputsList = [
+    [0.0],
+    [1.0],
+    [1.0],
+    [0.0],
+    [1.0],
+    [0.0],
+    [0.0],
+    [0.0],
+  ];
+
+  /// This function will calculate a fitness score after guessing with every
+  /// input within [LogicalXORFitnessService.logicalInputsList] on the input
+  /// [neuralNetwork].
+  @override
+  Future<double> gennScoringFunction({
+    required GENNNeuralNetwork neuralNetwork,
+  }) async {
+    // Collect all the guesses from this NeuralNetwork
+    final guesses = getNeuralNetworkGuesses(neuralNetwork: neuralNetwork);
+
+    // Declare a variable to store the sum of all errors
+    var errorSum = 0.0;
+
+    // Cycle through each guess to check its validity
+    for (int i = 0; i < guesses.length; i++) {
+      // Calculate the error from this guess
+      final error = (targetOutputsList[i][0] - guesses[i][0]).abs();
+
+      // Add this error to the errorSum
+      errorSum += error;
+    }
+
+    // Calculate the difference between a perfect score (8) and the total
+    // errors. A perfect score would mean zero errors with 8 correct answers,
+    // meaning a perfect score would be 8.
+    final diff = logicalInputsList.length - errorSum;
+
+    // To make the better performing Entities stand out more in this population,
+    // use the following equation to calculate the FitnessScore.
+    //
+    // 4 to the power of diff
+    return pow(4, diff).toDouble();
+  }
+
+  /// Returns the list of guesses (or outputs) from the input [neuralNetwork]
+  /// based on the standard set of inputs,
+  /// [LogicalXORFitnessService.logicalInputsList].
+  List<List<double>> getNeuralNetworkGuesses({
+    required GENNNeuralNetwork neuralNetwork,
+  }) {
+    // Declare a list of guesses
+    List<List<double>> guesses = [];
+
+    // Cycle through each input
+    for (int i = 0; i < logicalInputsList.length; i++) {
+      // Declare this run's set of inputs
+      final inputs = logicalInputsList[i];
+
+      // Make a guess using the NeuralNetwork
+      final guess = neuralNetwork.guess(inputs: inputs);
+
+      // Add this guess to the list of guesses
+      guesses.add(guess);
+    }
+
+    // Return the list of guesses
+    return guesses;
+  }
+}
